+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 
 * Added `applies_to_non_plan_charges` attribute to `Coupon`
+* Adding `gateway_error_code` to `Transaction`
 
 <a name="v2.4.5"></a>
 ## v2.4.5 (2015-7-31)