+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Adding `gateway_error_code` to `Transaction`
 * Adding `redemption_resource` to `Coupon`
 * Added `max_redemptions_per_account` attribute to `Coupon`
+* Added `redemptions` attribute to `Subscription`
 
 <a name="v2.4.5"></a>
 ## v2.4.5 (2015-7-31)