+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 
 * Adding `gateway_error_code` to `Transaction`
+* Added `setup_fee_accounting_code` attribute to `Plan`
 
 <a name="v2.4.5"></a>
 ## v2.4.5 (2015-7-31)